# backend/src/main/resources/contracts/autospec-backend-v1.openapi.yaml
openapi: 3.0.3
info:
  title: AutoSpec Backend API
  version: v1
  description: Spring Boot API contract for AutoSpec project orchestration, artifacts, exports, and observability.
servers:
  - url: /
security:
  - SessionToken: []
paths:
...
  /api/projects:
    get:
      summary: List projects visible to the session user.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Project history visible to the session user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a project from a natural-language requirement.
      parameters:
        - $ref: '#/components/parameters/SessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Created project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/projects/{projectId}/exports:
    get:
      summary: List durable export files without large content payloads.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Export metadata list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExportFileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/projects/{projectId}/exports/{exportFileId}:
    get:
      summary: Fetch a stored export file with content.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ExportFileId'
        - $ref: '#/components/parameters/SessionToken'
      responses:
        '200':
          description: Export file detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportFileDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/projects/{projectId}/export:
    post:
      summary: Render and persist a Markdown or PDF export.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SessionToken'
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [MARKDOWN, PDF]
            default: MARKDOWN
      responses:
        '200':
          description: Generated export payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    SessionToken:
      type: apiKey
      in: cookie
      name: AUTOSPEC_SESSION
  parameters:
    SessionToken:
      name: AUTOSPEC_SESSION
      in: cookie
      required: true
      schema:
        type: string
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ExportFileId:
      name: exportFileId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    PaginationLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    PaginationOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    Forbidden:
      description: Session user lacks the required project role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Project-scoped resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name, requirement]
      properties:
        name:
          type: string
          minLength: 1
        requirement:
          type: string
          minLength: 1
    CreateProjectResponse:
      type: object
      required: [projectId, status]
      properties:
        projectId:
          type: integer
          format: int64
        status:
          type: string
    ProjectResponse:
      type: object
      required: [projectId, name, originalRequirement, status, createdAt, updatedAt]
      properties:
        projectId:
          type: integer
          format: int64
        name:
          type: string
        originalRequirement:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ExportResponse:
      type: object
      required: [format, content, fileName, mediaType, encoding]
      properties:
        format:
          type: string
        content:
          type: string
        fileName:
          type: string
        mediaType:
          type: string
        encoding:
          type: string
    ExportFileResponse:
      type: object
      required: [id, projectId, fileName, mediaType, encoding, createdAt]
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        fileName:
          type: string
        mediaType:
          type: string
        encoding:
          type: string
        createdAt:
          type: string
          format: date-time
    ExportFileDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ExportFileResponse'
        - type: object
          required: [content]
          properties:
            content:
              type: string
    ApiErrorResponse:
      type: object
      required: [code, status, message, path, timestamp, fieldErrors, details]
      properties:
        code:
          type: string
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        details:
          type: object
          additionalProperties:
            type: string